refactor: 상태 이넘 연산 수정
